+++ cmd/hello/helloserver/hello.tmpl.html
